Improve FieldConstant unit test.
- merge field_const_test and field_evaluate_constant_test
- remove tests of value method, use only new assemblation
